% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_source_ui.R
\name{api_token_inputs}
\alias{api_token_inputs}
\title{API token inputs}
\usage{
api_token_inputs(
  label = "UDS-2:",
  placeholder = "Enter REDCap API Token for UDS-2",
  inputId = "uds_api_token"
)
}
\arguments{
\item{label}{A single string. Optional.}

\item{placeholder}{A single string. Optional.}

\item{inputId}{A single string. Optional.}
}
\value{
A Shiny \code{fluidRow} UI element containing a labeled password input field.
}
\description{
API token inputs
}
\keyword{internal}
