% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_panda.R
\name{bio_tab_for_gt}
\alias{bio_tab_for_gt}
\title{Format a table for gt}
\usage{
bio_tab_for_gt(tab)
}
\arguments{
\item{tab}{A \code{data.table}.}
}
\value{
A \code{data.table} with modified column names and values, formatted for use with \code{gt::gt}.
Returns \code{NULL} if the input \code{tab} is \code{NULL}.
}
\description{
Takes a table as those in the entries of the list provided by \code{get_biomarker_data},
and formats this for "pretty printing" using \code{gt::gt}.
}
\keyword{internal}
