% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_panda.R
\name{bio_tab_to_gt}
\alias{bio_tab_to_gt}
\title{Transform a table to a gt object}
\usage{
bio_tab_to_gt(tab_for_gt)
}
\arguments{
\item{tab_for_gt}{A \code{data.table} or a \code{list} of \code{data.table}s.}
}
\value{
A \code{gt::gt} table. If \code{tab_for_gt} is a list, the names of the list are used as
grouping variable for table.
}
\description{
A short description...
}
\keyword{internal}
