% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bio_tab_to_html_table.R
\name{bio_tab_to_html_table}
\alias{bio_tab_to_html_table}
\title{Convert biomarker data table to HTML table}
\usage{
bio_tab_to_html_table(
  tab_for_gt,
  densities,
  cuts,
  www_path = system.file("www", package = "ntrdWisconsin"),
  print_x = FALSE,
  tab_id = "csf",
  tab_header = "method",
  section_pattern = "^braak",
  section_label = "Braak Rating"
)
}
\arguments{
\item{tab_for_gt}{A list of \code{data.table}s, where each \code{data.table} is named after the Panda table that was queried to get the data.}

\item{densities}{A list of density data.}

\item{cuts}{A list or data structure containing cut-off values.}

\item{www_path}{Path to www/ folder that contains .css and .js files. Defaults to either inst/www or www subdir of NpsychAssessmentTool pacakge.}

\item{print_x}{For debugging}

\item{tab_id}{ID for the table, used to create unique cell IDs.}

\item{tab_header}{Header label for the table.}

\item{section_pattern}{Pattern to match on for defining a section. Default: \verb{"^braak}}

\item{section_label}{Label to show above section. Default: \code{"Braak Rating"}}
}
\value{
A \code{shiny::tagList} object representing the HTML table. Will error if \code{tab_for_gt} is not
a list of \code{data.table}s as expected.
}
\description{
Converts a list of biomarker \code{data.table}s (as returned by \code{\link[=bio_tab_for_gt]{bio_tab_for_gt()}})
into an interactive HTML table with inline density plots, threshold indicators,
and tooltip info icons for use in a Shiny application.
}
\seealso{
\code{\link[=density_plot]{density_plot()}}, \code{\link[=create_thresholds_table]{create_thresholds_table()}}
}
