% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomarker_thresholds.R
\docType{data}
\name{biomarker_thresholds}
\alias{biomarker_thresholds}
\title{Rules for Categorizing Biomarkers}
\format{
List of lists of lists such that \code{biomarker_thresholds[[table_name]][[biomarker]]} gives a list with entries \code{source} (list with \code{description} indicating where the categorization is described, and \code{link}), and \code{thresholds} (a \code{data.frame} with columns \code{label}, \code{min}, \code{max}, and \code{bin} describing categories).
}
\usage{
biomarker_thresholds
}
\description{
Rules for Categorizing Biomarkers
}
\keyword{datasets}
