% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npsych_scores-NACCGDS.R
\name{CESDTOTAL}
\alias{CESDTOTAL}
\title{CES-D Test Scores}
\usage{
CESDTOTAL(x = numeric())
}
\arguments{
\item{x}{Numeric scores.}
}
\value{
An object of class \code{CESDTOTAL}.
}
\description{
Create a \code{CESDTOTAL} object to hold CES-D Total scores
}
