% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_biomarker_data.R
\name{clean_biomarker_data}
\alias{clean_biomarker_data}
\title{Clean biomarker data}
\usage{
clean_biomarker_data(as_df, table_name, query)
}
\arguments{
\item{as_df}{A data.table of biomarker data. Should be the result of \code{jsonlite::fromJSON} applied to the result of API call.}

\item{table_name}{A single string indicating the name of the table from which the data originated.}

\item{query}{List obtained from calling \code{jsonlite::fromJSON} on the API query}
}
\value{
A cleaned \code{data.table} of biomarker data. The function may return an
error message (a string with class \code{"error-message"}) if no visits
could be matched to dates or if no non-missing biomarker data is found.
}
\description{
Clean up biomarker data from Panda.
}
