% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_threshold_table.R
\name{count_decimal_places}
\alias{count_decimal_places}
\title{Count decimal places}
\usage{
count_decimal_places(x)
}
\arguments{
\item{x}{A numeric value.}
}
\value{
A single integer representing the number of decimal places in \code{x}.
}
\description{
Get number of decimal places in a number.
}
\examples{
count_decimal_places(3.14)
count_decimal_places(0.00025)

}
