% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_source_server.R
\name{data_source_server}
\alias{data_source_server}
\title{Data source server for wadrc_source}
\arguments{
\item{source}{An S7 \code{wadrc_source} object.}

\item{id}{A single string, the Shiny module ID.}
}
\value{
A \code{shiny::moduleServer} function that returns a list containing:
\itemize{
\item \code{params}: A reactive list of API tokens, requiring \code{uds2_api_token},
\code{uds3_api_token}, and \code{uds4_api_token} to be present.
\item \code{restore}: A function to update the API token input fields.
\item \code{extras}: A \code{shiny::reactiveValues} object holding \code{all_values} (data
fetched using \code{panda_api_token}) and \code{panda_api_token}.
\item \code{session}: The Shiny session object.
}
}
\description{
A short description...
}
