% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_source_ui.R
\name{data_source_ui}
\alias{data_source_ui}
\title{Data source UI for WADRC source}
\arguments{
\item{source}{An S7 \code{wadrc_source} object.}

\item{ns}{A Shiny namespace function.}
}
\value{
A \code{shiny::tagList} containing UI elements for configuring REDCap API tokens and a Panda API Key.
}
\description{
Data source UI for WADRC source
}
