% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomarker_density_plot.R
\name{density_plot}
\alias{density_plot}
\title{Plot density}
\usage{
density_plot(
  obs = 30,
  dens = all_densities[[2]]$pTau_raw,
  cuts = all_cuts[[2]][name == "pTau", list(color, min_obs, max_obs)],
  height = 100,
  width = 400,
  new_id = NULL
)
}
\arguments{
\item{obs}{A numeric value. Optional.}

\item{dens}{A list or data frame containing density \code{x} and \code{y} values.}

\item{cuts}{A numeric vector of cut points.}

\item{height}{A single numeric for the plot height. Optional.}

\item{width}{A single numeric for the plot width. Optional.}

\item{new_id}{String to use for elementId of plotly container to avoid randomly assigned IDs.}
}
\value{
A \code{plotly} object. Returns \code{NULL} invisibly if \code{obs} is \code{NA}.
}
\description{
A short description...
}
