% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_panda.R
\name{get_biomarker_data}
\alias{get_biomarker_data}
\title{Fetch all biomarker values from Panda}
\usage{
get_biomarker_data(
  api_key = getOption("panda_api_key"),
  base_query_file = "inst/json/plasma.json",
  adrc_ptids = NULL
)
}
\arguments{
\item{api_key}{A single string. Panda API key.}

\item{base_query_file}{Path to the JSON query template file.}

\item{adrc_ptids}{Vector of ID's of ADRC participants for which biomarker
data should be pulled. If NULL (default), data for all participants are
pulled.}
}
\value{
A named list of \code{data.table}s, one per biomarker table.
}
\description{
Queries the Panda API for all biomarker data across all tables (excluding
Participants, Appointments, and Visual Rating tables). Used to compute
population-level density estimates and threshold cut-offs.
}
\keyword{internal}
