% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomarker_ratios.R
\name{lumipulse_ABeta42_ABeta40_ratio}
\alias{lumipulse_ABeta42_ABeta40_ratio}
\title{Lumipulse ABeta42/ABeta40 ratio}
\usage{
lumipulse_ABeta42_ABeta40_ratio(ABeta40, ABeta42)
}
\arguments{
\item{ABeta40}{A numeric value or vector.}

\item{ABeta42}{A numeric value or vector.}
}
\value{
A numeric vector representing the ABeta42/ABeta40 ratio. Values for \code{ABeta42} are
capped at 2200. Ratios will be \code{NA} if \code{ABeta42 < 38}, \code{ABeta40 < 158}, or \code{ABeta40 > 28450}.
For details, see \href{https://www.accessdata.fda.gov/cdrh_docs/pdf20/DEN200072.pdf}{this FDA document}.
}
\description{
Take vectors of ABeta40 and ABeta42 values and return the ratio.
}
\examples{
lumipulse_ABeta42_ABeta40_ratio(ABeta40 = c(5000, 100, 30000), ABeta42 = c(800, 40, 2500))

}
\seealso{
\code{\link[=lumipulse_pTau217_ABeta42_ratio]{lumipulse_pTau217_ABeta42_ratio()}}, \code{\link[=categorize_ratio]{categorize_ratio()}}
}
