% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomarker_ratios.R
\name{lumipulse_pTau217_ABeta42_ratio}
\alias{lumipulse_pTau217_ABeta42_ratio}
\title{Calculate Lumipulse pTau217/ABeta42 ratio}
\usage{
lumipulse_pTau217_ABeta42_ratio(pTau217, ABeta42)
}
\arguments{
\item{pTau217}{A numeric vector of pTau217 values.}

\item{ABeta42}{A numeric vector of ABeta42 values.}
}
\value{
A numeric vector representing the adjusted and rounded pTau217/ABeta42 ratio.
Values in \code{pTau217} are capped at 10000. Values in \code{ABeta42} are capped at 500
under specific conditions. Certain combinations of input values will result
in \code{NA} or \code{0.047} in the returned ratio. For more details, see \href{https://www.accessdata.fda.gov/cdrh_docs/reviews/K242706.pdf}{this FDA document}
}
\description{
Take vectors of pTau217 and ABeta42 values and return the ratio according to FDA guidance.
}
