% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nacc_to_wadrc_uds3.R
\docType{data}
\name{nacc_to_wadrc_uds3}
\alias{nacc_to_wadrc_uds3}
\title{Translate NACC variable names to WADRC variable names (UDS-3)}
\format{
A named character vector where names are NACC variable names and
values are the corresponding WADRC variable names in the UDS-3 database.
}
\usage{
nacc_to_wadrc_uds3
}
\description{
Translate NACC variable names to WADRC variable names (UDS-3)
}
\seealso{
\link{wadrc_uds3_to_nacc}, \code{\link[=wadrc_data_prep]{wadrc_data_prep()}}
}
\keyword{datasets}
