% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nacc_to_wadrc_uds4.R
\docType{data}
\name{nacc_to_wadrc_uds4}
\alias{nacc_to_wadrc_uds4}
\title{Translate NACC variable names to WADRC variable names (UDS-4)}
\format{
A named character vector where names are NACC variable names and
values are the corresponding WADRC variable names in the UDS-4 database.
}
\usage{
nacc_to_wadrc_uds4
}
\description{
Translate NACC variable names to WADRC variable names (UDS-4)
}
\seealso{
\link{wadrc_uds4_to_nacc}, \code{\link[=wadrc_data_prep]{wadrc_data_prep()}}
}
\keyword{datasets}
