% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_load.R
\name{pull_redcap_data}
\alias{pull_redcap_data}
\title{Pull REDCap data}
\usage{
pull_redcap_data(token, fields, uds)
}
\arguments{
\item{token}{A single string, the REDCap API token.}

\item{fields}{A character vector of field names to pull.}

\item{uds}{A single string or number representing the UDS version.}
}
\value{
A prepared data table from \code{wadrc_data_prep()}. If the data pull fails,
it returns \code{NULL} and will show a notification in a Shiny app or emit a warning.
}
\description{
Pulls data from a WADRC REDCap database using the provided API token,
then prepares it via \code{\link[=wadrc_data_prep]{wadrc_data_prep()}}.
}
\seealso{
\code{\link[=wadrc_source]{wadrc_source()}}, \link{wadrc_uds3_redcap_fields}
}
