% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readOldSavedFile.R
\name{readOldSavedFile}
\alias{readOldSavedFile}
\title{Read old saved file}
\usage{
readOldSavedFile()
}
\value{
A Shiny app object. When run, the app allows users to upload a file and enter a password.
Upon clicking "Load", it attempts to decrypt the file using \code{safer::retrieve_object()}
and display a table of extracted \code{source} and \code{api} authentication details. The app will
halt with an error if the file is invalid or the password is incorrect.
}
\description{
Small shiny app to retrieve API tokens from previously saved data source .bin file.
}
