% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/respval_behavioral_data.R
\name{respval_behavioral_data}
\alias{respval_behavioral_data}
\title{Behavioral and response validity data}
\usage{
respval_behavioral_data(dat)
}
\arguments{
\item{dat}{A data.table containing NACC behavioral and response validity data.}
}
\value{
A data.table with combined and processed behavioral observation and response validity data.
}
\description{
A short description...
}
