% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wadrc_data_prep.R
\name{return_single}
\alias{return_single}
\title{Return a single unique non-NA value}
\usage{
return_single(x, if_multiple = NA)
}
\arguments{
\item{x}{A vector.}

\item{if_multiple}{Value to return when multiple unique non-\code{NA} values
exist. Defaults to \code{NA}.}
}
\value{
A single value.
}
\description{
Given a vector, returns the single unique non-\code{NA} value if one exists.
If all values are \code{NA}, returns \code{NA}. If multiple unique non-\code{NA} values
exist, returns \code{if_multiple}.
}
\keyword{internal}
