% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_braak.R
\name{summarize_braak}
\alias{summarize_braak}
\title{Summarize Braak Stages into Positivity Threshold}
\usage{
summarize_braak(braak_1, braak_2, braak_3, braak_4, braak_5, braak_6)
}
\arguments{
\item{braak_1}{numeric vector of 0's or 1's.}

\item{braak_2}{numeric vector of 0's or 1's.}

\item{braak_3}{numeric vector of 0's or 1's.}

\item{braak_4}{numeric vector of 0's or 1's.}

\item{braak_5}{numeric vector of 0's or 1's.}

\item{braak_6}{numeric vector of 0's or 1's.}
}
\description{
Takes (vectors) of Braak stages I-VI and returns the corresponding positivity threshold.
}
