% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wadrc_data_prep.R
\name{wadrc_data_prep}
\alias{wadrc_data_prep}
\title{Prepare WADRC UDS Data}
\usage{
wadrc_data_prep(adrc_data, uds = c("uds2", "uds3", "uds4"))
}
\arguments{
\item{adrc_data}{Data as gotten from WADRC REDCap database.
Could be a complete download to .csv file, then read into R,
or pulled directly from REDCap using the \code{REDCapR} package.}

\item{uds}{Specify if data are from UDS-2, UDS-3, or UDS-4 database.}
}
\description{
Renames columns from WADRC REDCap variable names to NACC variable names,
fills missing data within and across visits, and returns a cleaned
\code{data.table} ready for downstream use.
}
\seealso{
\code{\link[=pull_redcap_data]{pull_redcap_data()}}
}
\keyword{internal}
