% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wadrc_source.R
\name{wadrc_source}
\alias{wadrc_source}
\title{Wisconsin ADRC data source}
\usage{
wadrc_source()
}
\value{
An S7 object of class \code{wadrc_source}, inheriting from \code{data_source}, with
\code{name = "Wisconsin ADRC"} and \code{id = "wadrc_redcap"}.
}
\description{
Wisconsin ADRC data source
}
