% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wadrc_uds3_to_nacc.R
\docType{data}
\name{wadrc_uds3_to_nacc}
\alias{wadrc_uds3_to_nacc}
\title{Translate WADRC UDS-3 variable names to NACC variable names}
\format{
A named character vector where names are WADRC variable names from
the UDS-3 REDCap database and values are the corresponding NACC variable names.
}
\usage{
wadrc_uds3_to_nacc
}
\description{
Translate WADRC UDS-3 variable names to NACC variable names
}
\seealso{
\link{nacc_to_wadrc_uds3}, \code{\link[=wadrc_data_prep]{wadrc_data_prep()}}
}
\keyword{datasets}
