% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extensionModule.R
\name{extension_ui}
\alias{extension_ui}
\alias{extension_server}
\alias{extension_app}
\title{Biomarker UI}
\usage{
extension_ui(id = "biomarker-tables")

extension_server(
  id = "biomarker-table",
  ptid,
  extras,
  base_query_file = system.file("json/panda_template.json", package = "ntrdWisconsin")
)

extension_app(
  ptid,
  biomarker_api,
  all_values,
  base_query_file = system.file("json/panda_template.json", package = "ntrdWisconsin"),
  testing = FALSE
)
}
\arguments{
\item{id}{A string used to namespace the module.}

\item{ptid}{ADRC ptid. A character vector with patient ID's that can be chosen.}

\item{biomarker_api}{Panda API key.}

\item{all_values}{A reactive value}
}
\value{
A UI definition.

NULL.

A \code{shinyApp} object.
}
\description{
A short description...

A short description...

A short description...
}
\keyword{internal}
