% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_threshold_table.R
\name{create_thresholds_table}
\alias{create_thresholds_table}
\title{Create thresholds table}
\usage{
create_thresholds_table(thres)
}
\arguments{
\item{thres}{A data frame or tibble with \code{min}, \code{max}, and \code{label} columns.}
}
\value{
A \code{shiny::HTML} object representing an HTML table of the thresholds.
}
\description{
Create HTML table to include in tooltip for info column of bio_tab_to_html_table.
}
