% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_load.R
\name{data_load}
\alias{data_load}
\title{Load WADRC data}
\arguments{
\item{source}{An object of class \code{wadrc_source}.}

\item{uds2_api_token}{A single string.}

\item{uds3_api_token}{A single string.}

\item{uds4_api_token}{A single string.}

\item{...}{Currently unused; must be empty.}
}
\value{
An object of class \code{data_nacc} if data is successfully retrieved and combined.
If no data is retrieved from any UDS, \code{NULL} is returned and an error notification is shown in a Shiny context.
Notifications are displayed within a Shiny application to indicate data loading progress and completion.
}
\description{
Pulls UDS-2, UDS-3, and UDS-4 data from the WADRC REDCap databases,
prepares each dataset, combines them, and computes derived neuropsych scores.
}
\seealso{
\code{\link[=pull_redcap_data]{pull_redcap_data()}}, \code{\link[=wadrc_source]{wadrc_source()}}
}
\keyword{internal}
