% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_panda.R
\name{get_biomarker_data}
\alias{get_biomarker_data}
\title{Get biomarker data}
\usage{
get_biomarker_data(
  adrc_ptid = "adrc00006",
  api_key,
  base_query_file = system.file("json/panda_template.json", package = "ntrdWisconsin")
)
}
\arguments{
\item{adrc_ptid}{A single string. ADRC participant id for which we want to pull biomarker data.}

\item{api_key}{A single string. API key for panda database.}

\item{base_query_file}{A single string. Optional.}
}
\value{
A list of \code{data.table}s containing biomarker data, with names corresponding to the tables from
Panda that was queried.
}
\description{
Queries the \href{https://panda.medicine.wisc.edu}{Panda} database to get biomarker data for
ADRC participants.
}
\keyword{internal}
