% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nacc_to_wadrc_uds2.R
\docType{data}
\name{nacc_to_wadrc_uds2}
\alias{nacc_to_wadrc_uds2}
\title{Translate NACC variable names to WADRC variable names (UDS-2)}
\format{
A named vector where names of entries correspond to NACC variable
names, and entries contain WADRC variable names as is in UDS-2 database.
}
\usage{
nacc_to_wadrc_uds2
}
\description{
Translate NACC variable names to WADRC variable names (UDS-2)
}
\keyword{datasets}
