% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntrdWisconsin-package.R
\docType{package}
\name{ntrdWisconsin-package}
\alias{ntrdWisconsin}
\alias{ntrdWisconsin-package}
\title{ntrdWisconsin: Neuropsychological Test Result Dashboards for ADRC Wisconsin}
\description{
Extension to `ntrd` for the Wisconsin ADRC to import neuropsychological data directly from REDCap and biomarker data from Panda.
}
\author{
\strong{Maintainer}: Ralph Møller Trane \email{rmtrane@gmail.com} (\href{https://orcid.org/0000-0001-7824-1785}{ORCID})

}
\keyword{internal}
