% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wadrc_uds3_to_nacc.R
\docType{data}
\name{wadrc_uds3_to_nacc}
\alias{wadrc_uds3_to_nacc}
\title{Translate WADRC UDS-3 variable names to NACC variable names}
\format{
A named vector where names of entries correspond to WADRC variable
names as they appear in the UDS-3 REDCap database, and entries contain NACC
variable names.
}
\usage{
wadrc_uds3_to_nacc
}
\description{
Translate WADRC UDS-3 variable names to NACC variable names
}
\keyword{datasets}
