% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wadrc_uds4_redcap_fields.R
\docType{data}
\name{wadrc_uds4_redcap_fields}
\alias{wadrc_uds4_redcap_fields}
\title{Fields to pull from WADRC UDS-4 REDCap}
\format{
Character vector with names of fields we pull from the WADRC UDS-4 REDCap database
}
\usage{
wadrc_uds4_redcap_fields
}
\description{
Fields to pull from WADRC UDS-4 REDCap
}
\keyword{datasets}
